# postman/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shop-offices": {
      "post": {
        "operationId": "ShopOfficesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopOfficeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add new a shop office",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOfficeEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add new a shop office",
        "tags": [
          "ShopOffices"
        ]
      },
      "get": {
        "operationId": "ShopOfficesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": " shop office list found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShopOfficeEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get shop office list",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{id}": {
      "get": {
        "operationId": "ShopOfficesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office found!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOfficeEntity"
                }
              }
            }
          }
        },
        "summary": "Get shop office detail by ID",
        "tags": [
          "ShopOffices"
        ]
      },
      "patch": {
        "operationId": "ShopOfficesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShopOfficeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a shop office",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOfficeEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a shop office",
        "tags": [
          "ShopOffices"
        ]
      },
      "delete": {
        "operationId": "ShopOfficesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a shop office",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOfficeEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a shop office",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{id}/manager-list": {
      "get": {
        "operationId": "ShopOfficesController_findAllManagersOfShopOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office\"s manager list found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get shop office\"s manager list by ID shop office",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{id}/address": {
      "get": {
        "operationId": "ShopOfficesController_findAddressOfShopOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office\"s address found!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressEntity"
                }
              }
            }
          }
        },
        "summary": "Get shop office\"s address by ID shop office",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{id}/product-list": {
      "get": {
        "operationId": "ShopOfficesController_findAllProductsOfShopOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office\"s products found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get shop office\"s products by ID shop office",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{id}/category-list": {
      "get": {
        "operationId": "ShopOfficesController_findAllCategoryOfShopOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office\"s categories found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get shop office\"s categories by ID shop office",
        "tags": [
          "ShopOffices"
        ]
      }
    },
    "/shop-offices/{shopId}/category/{categoryId}/product-list": {
      "get": {
        "operationId": "ShopOfficesController_findAllProductsOfCategoryOfShopOffice",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "shop office\"s products of category found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get shop office\"s products of one category by ID shop office and ID category",
        "tags": [
          "ShopOffices"
        ]
      }
    }
  },
  "info": {
    "title": "BE Ecomerce Shop",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1,
            "description": "User ID"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "description": "Gender"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email address"
          },
          "phone": {
            "type": "string",
            "example": "0123456789",
            "description": "Phone number"
          },
          "username": {
            "type": "string",
            "example": "johndoe",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "example": "hashedpassword",
            "description": "Password hash"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "OPERATOR"
            ],
            "description": "User role"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Account active status"
          },
          "codeActive": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Activation code"
          },
          "codeActiveExpire": {
            "format": "date-time",
            "type": "string",
            "example": "2024-12-31T23:59:59Z",
            "description": "Activation code expiry"
          },
          "points": {
            "type": "number",
            "example": 100,
            "description": "Loyalty points"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00Z",
            "description": "Created at"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00Z",
            "description": "Updated at"
          },
          "staffCode": {
            "type": "string",
            "example": "STF001",
            "description": "Staff code"
          },
          "isAdmin": {
            "type": "boolean",
            "example": false,
            "description": "Is admin"
          },
          "shopOfficeId": {
            "format": "int64",
            "type": "integer",
            "example": 1,
            "description": "Shop office ID"
          },
          "loyaltyCard": {
            "type": "string",
            "example": "LOYAL123",
            "description": "Loyalty card number"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "password",
          "role",
          "isActive",
          "codeActive",
          "codeActiveExpire",
          "points",
          "createdAt",
          "updatedAt",
          "isAdmin"
        ]
      },
...
      "AddressEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "userId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "street": {
            "type": "string",
            "example": "123 Main Street"
          },
          "ward": {
            "type": "string",
            "example": "Ward 1"
          },
          "district": {
            "type": "string",
            "example": "District 1"
          },
          "province": {
            "type": "string",
            "example": "Ho Chi Minh City"
          },
          "zipCode": {
            "type": "string",
            "example": "70000"
          },
          "country": {
            "type": "string",
            "example": "Vietnam"
          },
          "shopOfficeId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "street",
          "ward",
          "district",
          "province",
          "zipCode",
          "country",
          "shopOfficeId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ShopOfficeEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "shopName": {
            "type": "string",
            "example": "Main Store"
          }
        },
        "required": [
          "id",
          "shopName"
        ]
      },
...
      "ProductEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Product Name"
          },
          "description": {
            "type": "object",
            "example": "Product description",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "stockKeepingUnit": {
            "type": "string",
            "example": "SKU123456"
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "categoryId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "voucherId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "shopOfficeId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "stockKeepingUnit",
          "stock",
          "categoryId",
          "createByUserId",
          "voucherId",
          "shopOfficeId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CategoryEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "description": {
            "type": "object",
            "example": "Electronic devices and accessories",
            "nullable": true
          },
          "parentId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "voucherId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "shopOfficeId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "parentId",
          "createByUserId",
          "voucherId",
          "shopOfficeId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateShopOfficeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateShopOfficeDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}